% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArdecheRiver}
\alias{ArdecheRiver}
\title{Ardèche River Streamflow}
\format{
A data frame with 731 rows and 3 variables:
\describe{
  \item{date}{date}
  \item{obs}{Observed streamflow (mm)}
  \item{sim}{Simulated streamflow (mm)}
}
}
\usage{
ArdecheRiver
}
\description{
Two years of observed and simulated streamflow for the Ardèche River at Sauze-St-Martin hydrometric station.
See https://en.wikipedia.org/wiki/Ardèche_(river) for a description of the river.
Simulated streamflow are from the GR4J hydrological model, available through
the airGR package (https://hydrogr.github.io/airGR/).
}
\keyword{datasets}
