% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTransform}
\alias{getTransform}
\title{Predictor Transformation}
\usage{
getTransform(values, trans = c("none", "normal", "uniform"))
}
\arguments{
\item{values}{numeric vector, values used to establish the transformation. Typically,
the predictor values that were used for calibration}

\item{trans}{character, the requested transformation, one of 'none', 'normal' or 'uniform'}
}
\value{
a function that can be applied to some predictor values
    (typically, the ones used for prediction)
}
\description{
Get the transformation function that can be applied to a predictor
}
\examples{
n=1000
values=rlnorm(n)
f=getTransform(values,'uniform')
plot(values,f(values))
hist(f(values))
x=rnorm(n,0,10)
plot(x);points(values,col='red')
plot(x,f(x))
hist(f(x))
}
