% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialRegression.R
\name{HermitePolynomials}
\alias{HermitePolynomials}
\title{Hermite Polynomials}
\usage{
HermitePolynomials(x, n = 3)
}
\arguments{
\item{x}{numeric vector, values at which polynomials are computed.}

\item{n}{Integer, maximum polynomial degree.}
}
\value{
A data frame with n+1 columns, where column k is the Hermite polynomial
    with degree k-1.
}
\description{
Compute Hermite Polynomials from degree 0 up to degree n
}
\examples{
x=seq(-3,3,0.01)
DF=HermitePolynomials(x,5)
plot(NA,xlab='x',ylab='Hn(x)',xlim=range(x),ylim=range(DF))
for(i in 1:NCOL(DF)){lines(x,DF[,i],col=i)}
}
