% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialRegression.R
\name{LegendrePolynomials}
\alias{LegendrePolynomials}
\title{Legendre Polynomials}
\usage{
LegendrePolynomials(x, n = 3)
}
\arguments{
\item{x}{numeric vector, values at which polynomials are computed.}

\item{n}{Integer, maximum polynomial degree.}
}
\value{
A data frame with n+1 columns, where column k is the Legendre polynomial
    with degree k-1.
}
\description{
Compute Legendre Polynomials from degree 0 up to degree n
}
\examples{
x=seq(-1,1,0.01)
DF=LegendrePolynomials(x,5)
plot(NA,xlab='x',ylab='Ln(x)',xlim=range(x),ylim=range(DF))
for(i in 1:NCOL(DF)){lines(x,DF[,i],col=i)}
}
