% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialRegression.R
\name{plot.polyRegPred}
\alias{plot.polyRegPred}
\title{Plot Polynomial Regression Prediction}
\usage{
\method{plot}{polyRegPred}(
  x,
  predictorType = c("scaled", "raw", "transformed"),
  intervalLevel = 0.9,
  ...
)
}
\arguments{
\item{x}{object of class polyRegPred, resulting from a call to function [predict.polyRegFit()]}

\item{predictorType}{character string, one of 'scaled' (show rescaled predictor), 'raw' (show raw predictor)
or 'transformed' (show transformed predictor, only available when x is of class anaResPred).}

\item{intervalLevel}{numeric in (0,1), level of the uncertainty intervals.}

\item{...}{further plotting arguments passed to or from other methods.}
}
\value{
A list of 3 ggplots:
\enumerate{
  \item 'mu': for each mu predictor, a 2-panel plot with the predictor on the x-axis,
      and on the y-axis: estimated mu, predicted values
  \item 'sigma': same as previously but for sigma
  \item 'phi': same as previously but for phi
}
}
\description{
Plot prediction from a fitted polynomial regression.
}
\examples{
n=366
predictor=rnorm(n)
y=rnorm(n,mean=2*sin(predictor),sd=0.1*(abs(predictor)+0.5))
reg=polynomialRegression(y,predictor)
newpred=rnorm(100,1.5,1)
res=predict(reg,newdata=list(mu=newpred,sigma=newpred,phi=newpred))
plot(res)
}
