% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialRegression.R
\name{polynomialRegression}
\alias{polynomialRegression}
\title{Polynomial Regression}
\usage{
polynomialRegression(
  y,
  x_mu,
  x_sigma = x_mu,
  x_phi = x_mu,
  polyFunk = LegendrePolynomials,
  deg_mu = rep(3, NCOL(x_mu)),
  deg_sigma = rep(2, NCOL(x_sigma)),
  deg_phi = rep(0, NCOL(x_phi)),
  link_mu = identity,
  link_sigma = exp,
  link_phi = rescaledLogistic,
  useApproach2 = FALSE,
  optim_control = list(maxit = 1e+05),
  optim_method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent")
)
}
\arguments{
\item{y}{numeric vector, predictand.}

\item{x_mu}{numeric vector, matrix or data frame, predictors for the mean mu.}

\item{x_sigma}{numeric vector, matrix or data frame, predictors for the sdev sigma.}

\item{x_phi}{numeric vector, matrix or data frame, predictors for the autocorrelation phi.}

\item{polyFunk}{function, function used to compute predictor polynomials.}

\item{deg_mu}{Integer vector, size NCOL(x_mu). Polynomial degree of each predictor
used to compute the mean. Trick: passing a single value as a character will force
mu to remain equal to this value.}

\item{deg_sigma}{integer vector, size NCOL(x_sigma). Polynomial degree of each predictor
used to compute the standard deviation. Trick: passing a single value as a character will force
sigma to remain equal to this value.}

\item{deg_phi}{integer vector, size NCOL(x_phi). Polynomial degree of each predictor
used to compute the autocorrelation. Trick: passing a single value as a character will force
phi to remain equal to this value.}

\item{link_mu}{function, link function used for the mean mu.}

\item{link_sigma}{function, link function used for the standard deviation sigma.}

\item{link_phi}{function, link function used for the autocorrelation phi.}

\item{useApproach2}{logical, if TRUE "approach 2" is used (based on center-scaled residuals), otherwise approach 1 is used.}

\item{optim_control}{list, list of controls to be passed to the optimizer, see ?optim.}

\item{optim_method}{character string, optimization method, see ?optim.}
}
\value{
An object of class 'PolyRegFit', which is a list with the following components:
  \item{theta}{numeric vector, estimated parameter vector}
  \item{coeff}{list of size 3, subvector of theta corresponding to
      the polynomial coefficients for mu, sigma and phi}
  \item{pred}{list of size 3, estimated mean mu, sdev sigma and autocorrelation phi (same size as y)}
  \item{residuals}{numeric vector, standardized (centered-scaled) residuals (same size as y)}
  \item{innovations}{numeric vector, innovations (i.e. whitened centered-scaled residuals, same size as y)}
  \item{poly}{list of size 3, polynomials used to compute mu and sigma}
  \item{xscaled}{list of size 3, scaled predictors for mu, sigma and phi, with scaling depending
      on polyFunk (Hermite and Standard: center-scale, Legendre: send between -1 and 1)}
  \item{scaler}{list of size 3, each element being a list of functions:
      scaling function used for each predictor in x_mu, x_sigma and x_phi}
  \item{scalerInv}{list of size 3, each element being a list of functions:
      inverse scaling function for each predictor in x_mu, x_sigma and x_phi}
  \item{scalerPar}{list of size 2, parameters of each scaling function}
  \item{optim}{list, optimisation results returned by [optim()]}
  \item{y}{numeric vector, predictand}
  \item{x}{list of size 3, predictors for mu, sigma and phi}
  \item{deg}{list of size 3, polynomial degree associated to each predictor for computing mu, sigma and phi}
  \item{polyFunk}{function, function to compute polynomials}
  \item{useApproach2}{logical, if TRUE "approach 2" is used (based on center-scaled residuals), otherwise approach 1 is used}
  \item{link_mu}{function, link function used for the mean mu}
  \item{link_sigma}{function, link function used for the standard deviation sigma}
  \item{link_phi}{function, link function used for the autocorrelation phi}
}
\description{
Fit a polynomial regression for the mean mu, the standard deviation sigma
and the autocorrelation phi of an AR(1) Gaussian distribution
}
\examples{
n=366
predictor=rnorm(n)
y=rnorm(n,mean=2*sin(predictor),sd=0.1*(abs(predictor)+0.5))
plot(predictor,y)
reg=polynomialRegression(y,predictor)
plot(reg)
}
