% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescaledLogistic}
\alias{rescaledLogistic}
\title{Rescaled Logistic Function}
\usage{
rescaledLogistic(x)
}
\arguments{
\item{x}{numeric, values at which the function is evaluated}
}
\value{
a numeric equal to L(x)
}
\description{
A logistic function rescaled between -1 and 1, L(x)=(exp(x)-1)/(exp(x)+1)
}
\examples{
x=seq(-10,10,0.01)
y=rescaledLogistic(x)
plot(x,y)
}
