% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertaintyEstimation.R
\name{analyseResiduals}
\alias{analyseResiduals}
\title{Residuals Analyzer}
\usage{
analyseResiduals(
  obs,
  sim,
  x_mu = data.frame(sim = sim),
  x_sigma = data.frame(sim = sim),
  x_phi = data.frame(dsim = c(0, diff(sim))),
  polyFunk = LegendrePolynomials,
  trans = c("uniform", "normal", "none"),
  deg_mu = 5,
  deg_sigma = 5,
  deg_phi = 5,
  ...
)
}
\arguments{
\item{obs}{numeric vector, observed values.}

\item{sim}{numeric vector, simulated values.}

\item{x_mu}{numeric vector, matrix or data frame, predictors for the mean mu.}

\item{x_sigma}{numeric vector, matrix or data frame, predictors for the sdev sigma.}

\item{x_phi}{numeric vector, matrix or data frame, predictors for the autocorrelation phi.}

\item{polyFunk}{function, function used to compute predictor polynomials.}

\item{trans}{character string, predictor transformation, one of c('uniform','normal','none').}

\item{deg_mu}{Integer vector, size NCOL(x_mu). Polynomial degree of each predictor
used to compute the mean. Trick: passing a single value as a character will force
mu to remain equal to this value.}

\item{deg_sigma}{integer vector, size NCOL(x_sigma). Polynomial degree of each predictor
used to compute the standard deviation. Trick: passing a single value as a character will force
sigma to remain equal to this value.}

\item{deg_phi}{integer vector, size NCOL(x_phi). Polynomial degree of each predictor
used to compute the autocorrelation. Trick: passing a single value as a character will force
phi to remain equal to this value.}

\item{...}{additional arguments passed to function [polynomialRegression()]}
}
\value{
An object of class 'anaRes', which is an object of class 'polyRegFit'
    augmented with the following components:
  \item{obs}{numeric vector, observed values}
  \item{sim}{numeric vector, simulated values}
  \item{xtrans}{list of size 3, transformed predictors for mu, sigma and phi}
  \item{trans}{list of size 3, each element being a list of functions:
      transformation function used for each predictor (column) in x_mu, x_sigma and x_phi}
}
\description{
Analyse the residuals between obs and sim.
Residuals are assumed to be realizations from an AR(1) Gaussian process,
whose mean mu, standard deviation sigma and lag-1 autocorrelation phi
vary according to polynomial regressions with predictors.
}
\examples{
w=analyseResiduals(ArdecheRiver$obs,ArdecheRiver$sim)
plot(w)
}
