% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertaintyEstimation.R
\name{getUncertainty}
\alias{getUncertainty}
\title{Predictive Uncertainty}
\usage{
getUncertainty(
  object,
  sim = object$sim,
  x = list(mu = data.frame(sim = sim), sigma = data.frame(sim = sim), phi =
    data.frame(dsim = c(0, diff(sim)))),
  nrep = 1000,
  intervalLevel = 0.9,
  ...
)
}
\arguments{
\item{object}{object of class anaRes, resulting from a call to function [analyseResiduals()]}

\item{sim}{numeric vector, simulated values}

\item{x}{list of size 3 with names 'mu', 'sigma' and 'phi', predictors (matrix or data frame).}

\item{nrep}{integer, number of replications used to compute parametric and predictive
uncertainties.}

\item{intervalLevel}{numeric in (0,1), level of the uncertainty intervals.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'PredictiveU', which is an object of class 'polyRegPred'
    augmented with the following components:
  \item{sim}{numeric vector, simulated values}
  \item{spag}{numeric matrix, size nrep*length(sim): replications representing predictive uncertainty}
  \item{env}{data frame, with columns 'median' (predictive median) and 'low' + 'high' (predictive uncertainty envelope). }
}
\description{
Compute predictive uncertainty around simulated values, after having performed a residual analysis.
}
\examples{
w=analyseResiduals(ArdecheRiver$obs[1:365],ArdecheRiver$sim[1:365])
u=getUncertainty(w,sim=ArdecheRiver$sim[366:730])
plot(u,axisValues=ArdecheRiver$date[366:730])
}
