% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertaintyEstimation.R
\name{getUncertainty}
\alias{getUncertainty}
\title{Predictive Uncertainty}
\usage{
getUncertainty(
  object,
  sim = object$sim,
  x = list(mu = data.frame(sim = sim), sigma = data.frame(sim = sim), phi =
    data.frame(dsim = c(0, diff(sim)))),
  nsim = 1000,
  probs = c(0.05, 0.5, 0.95),
  ...
)
}
\arguments{
\item{object}{object of class anaRes, resulting from a call to function [analyseResiduals()]}

\item{sim}{numeric vector, simulated values}

\item{x}{list of size 3 with names 'mu', 'sigma' and 'phi', predictors (matrix or data frame).}

\item{nsim}{integer, number of replications used to compute parametric and predictive
uncertainties. Uncertainties are skipped if nsim<=0}

\item{probs}{numeric vector of length 3, probabilities used to compute the low/middle/high bounds of uncertainty envelops}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'PredictiveU', which is a list with the following components:
  \item{XXX}{XXX}
}
\description{
Compute predictive uncertainty around simulated values, after having performed a residual analysis.
}
\examples{
q=as.numeric(Nile)
obs=0.1*q^(3/5)+rnorm(length(q),sd=0.2)
sim=0.11*q^(3.1/5)
plot(sim,type='l');points(obs)
w=analyseResiduals(obs,sim)
u=getUncertainty(w)
plot(u)
}
