% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NSE}
\alias{NSE}
\title{Nash-Sutcliffe Efficiency}
\usage{
NSE(obs, sim)
}
\arguments{
\item{obs}{numeric vector, observed values}

\item{sim}{numeric vector, simulated values}
}
\value{
a numeric value equal to the NSE
}
\description{
Compute the Nash-Sutcliffe Efficiency (NSE)
}
\examples{
NSE(ArdecheRiver$obs,ArdecheRiver$sim)
plot(ArdecheRiver$obs);lines(ArdecheRiver$sim,col='red')
}
