% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertaintyEstimation.R
\name{plot.anaRes}
\alias{plot.anaRes}
\title{Plot Result of Residual Analysis}
\usage{
\method{plot}{anaRes}(
  x,
  predictorType = c("transformed", "raw", "scaled"),
  alpha = 0.25,
  allPlots = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class 'anaRes', resulting from a call to function [analyseResiduals()]}

\item{predictorType}{character string, one of 'transformed' (show transformed predictors),
'raw' (show raw predictors) or 'scaled' (show rescaled predictors).}

\item{alpha}{numeric in [0,1], transparency of uncertainty interval.}

\item{allPlots, }{boolean, produce all plots or only the obs vs. sim one?}

\item{...}{further plotting arguments passed to or from other methods.}
}
\value{
A list of 4 ggplots (if allPlots=FALSE, only the first one is returned):
\enumerate{
  \item 'ObsVsSim': predicted values (with uncertainties) and observations
  \item 'mu': for each mu predictor, a 3-panel plot with the predictor on the x-axis,
      and on the y-axis: estimated mu, predicted values and standardized residuals
  \item 'sigma': same as previously but for sigma
  \item 'phi': same as previously but for phi
}
}
\description{
Plot results of a residual analysis.
}
\examples{
w=analyseResiduals(ArdecheRiver$obs[1:366],ArdecheRiver$sim[1:366])
plot(w)
}
