% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialRegression.R
\name{plot.polyRegFit}
\alias{plot.polyRegFit}
\title{Plot Polynomial Regression}
\usage{
\method{plot}{polyRegFit}(
  x,
  predictorType = c("scaled", "raw", "transformed"),
  alpha = 0.25,
  intervalLevel = 0.9,
  ...
)
}
\arguments{
\item{x}{object of class polyRegFit, resulting from a call to function [polynomialRegression()]}

\item{predictorType}{character string, one of 'scaled' (show rescaled predictor), 'raw' (show raw predictor)
or 'transformed' (show transformed predictor, only available when x is of class anaRes).}

\item{alpha}{numeric in [0,1], transparency of uncertainty intervals.}

\item{intervalLevel}{numeric in (0,1), level of the uncertainty intervals.}

\item{...}{further plotting arguments passed to or from other methods.}
}
\value{
A list of 4 ggplots:
\enumerate{
  \item 'obs': predicted values (with uncertainties) and standardized residuals
  \item 'mu': for each mu predictor, a 3-panel plot with the predictor on the x-axis,
      and on the y-axis: estimated mu, predicted values and standardized residuals
  \item 'sigma': same as previously but for sigma
  \item 'phi': same as previously but for phi
}
}
\description{
Plot results of a polynomial regression.
}
\examples{
n=366
predictor=rnorm(n)
y=rnorm(n,mean=2*sin(predictor),sd=0.1*(abs(predictor)+0.5))
plot(predictor,y)
reg=polynomialRegression(y,predictor)
plot(reg)
}
