% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertaintyEstimation.R
\name{plot.predictiveUncertainty}
\alias{plot.predictiveUncertainty}
\title{Plot Predictive Uncertainty}
\usage{
\method{plot}{predictiveUncertainty}(
  x,
  axisValues = data.frame(Index = 1:NROW(x$sim)),
  labs = c(names(axisValues), "Prediction"),
  col = "red",
  alpha = 0.25,
  showRawSim = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class 'predictiveUncertainty', resulting from a call to function [getUncertainty()]}

\item{axisValues}{numeric or date vector, what to display on the x-axis.}

\item{labs}{character vector of size 2, x-axis and y-axis labels.}

\item{col}{color specification, color used for predictions.}

\item{alpha, }{numeric in (0,1), transparency of uncertainty interval.}

\item{showRawSim}{boolean, add raw simulations (before mu-correction) to plot ?}

\item{...}{further plotting arguments passed to or from other methods.}
}
\value{
A ggplot showing the simulated values + uncertainties
}
\description{
Plot predictive uncertainty estimated after a residual analysis.
}
\examples{
q=as.numeric(Nile)
obs=0.1*q^(3/5)+rnorm(length(q),sd=0.2)
sim=0.11*q^(3.1/5)
plot(sim,type='l');points(obs)
w=analyseResiduals(obs,sim)
u=getUncertainty(w)
plot(u)
}
