% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomialRegression.R
\name{predict.polyRegFit}
\alias{predict.polyRegFit}
\title{Polynomial Regression Prediction}
\usage{
\method{predict}{polyRegFit}(object, newdata = NULL, nsim = 1000, ...)
}
\arguments{
\item{object}{object of class polyRegFit, resulting from a call to function [polynomialRegression()].}

\item{newdata}{list of size 3 with names 'mu', 'sigma' and 'phi', predictors (matrix or data frame).
If NULL, the predictors used for calibration are used.}

\item{nsim}{integer, number of replications used to compute parametric and predictive
uncertainties. Uncertainties are skipped if nsim<=0}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'PolyRegPred', which is a list with the following components:
  \item{pred}{list of size 3, estimated mean mu, sdev sigma and autocorrelation phi}
  \item{replicates}{list of size 3, nsim replicates of mu, sigma and phi,
      representing parametric uncertainty (size nsim*NROW(newdata[['mu']]))}
  \item{predictive}{numeric matrix, nsim replicates of predictand values, representing
      total predictive uncertainty (size nsim*NROW(newdata[['mu']]))}
  \item{x}{list of size 3, predictors for mu, sigma and phi}
  \item{xscaled}{list of size 3, scaled predictors for mu, sigma and phi}
  \item{poly}{list of size 3, polynomials used to compute mu, sigma and phi}
  \item{C}{matrix, covariance matrix of inferred parameters (approximated using the Hessian)}
}
\description{
Predict from a polynomial regression for both the mean, the standard deviation
and the autocorrelation of a AR(1) Gaussian distribution.
}
\examples{
n=366
predictor=rnorm(n)
y=rnorm(n,mean=2*sin(predictor),sd=0.1*(abs(predictor)+0.5))
reg=polynomialRegression(y,predictor)
newpred=rnorm(100,1.5,1)
res=predict(reg,newdata=list(mu=newpred,sigma=newpred,phi=newpred))
plot(res)
}
